requirements.txt
numpy==1.19.2
opencv_python==4.6.0.66
Pillow==9.2.0
